Ajout des methodes CRUD DELETE et POST à l'album

+++ AlbumPhoto/src/controleur/Album.java
@@ -1,6 +1,7 @@
 package controleur;
 
 import java.io.IOException;
+import java.util.Date;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -10,11 +11,19 @@ import javax.servlet.http.HttpServletResponse;
 
 import dao.DAOFactory;
 import extra.Data;
+import modele.Personne;
 
 @WebServlet(urlPatterns = { "/Album", "/Album/*" })
 public class Album extends HttpServlet {
 	private static final long serialVersionUID = 1L;
-
+	public static final String MESSAGE_DELETE_200 = "L'album a été supprimé avec succès";
+	public static final String MESSAGE_DELETE_404 = "Une erreur est survenue. L'album n'existe pas.";
+	public static final String MESSAGE_POST_UTILISATEUR_NULL_404 = "Cet utilisateur n'existe pas.";
+	public static final String MESSAGE_POST_200 = "L'album a été créé";
+	public static final String MESSAGE_POST_UTILISATEUR_DIFF_404 = "Vous n'avez pas les droits pour modifier cet utilisateur.";
+	public static final String MESSAGE_POST_ATTRIBUTE_MISSING = "Le titre n'a pas été rempli.";
+	public static final String MESSAGE_DELETE_ATTRIBUTE_MISSING = "L'id n'est pas renseigné.";
+	
 	public Album() {
 		super();
 	}
@@ -52,55 +61,62 @@ public class Album extends HttpServlet {
 
 	@Override
 	protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		System.out.println("BiduleDELETE !!");
-		System.out.println(request.getParameter("idAlbum"));
-		/*
-		 * On récupére les paramètre en passant pas une fonction car dans un
-		 * méthode avec la méthode DELETE les paramètres ne sont pas stockés
-		 * dans une map
-		 */
-		// HashMap parameters = (HashMap)
-		// parseParameters(request.getInputStream());
-		// System.out.println("BiduleDELETE !!");
-		// System.out.println(parameters.get("idAlbum"));
-		//
-		// Récupére l'album passé en paramètre
-		modele.Album album = (modele.Album) DAOFactory.getInstance().getAlbumDao().read(Integer.getInteger(request.getParameter("idAlbum")));
-		System.out.println("Album : " + album);
-		// Si l'album existe
-		if (album != null) {
-			DAOFactory.getInstance().getAlbumDao().delete(album);
+		String idAlbum = request.getParameter("idAlbum");
+		if(idAlbum != ""){
+			// Récupére l'album passé en paramètre
+			modele.Album album = (modele.Album) DAOFactory.getInstance().getAlbumDao().read(Integer.getInteger(idAlbum));
+			
+			// Si l'album existe
+			if (album != null) {
+				DAOFactory.getInstance().getAlbumDao().delete(album);
+				request.setAttribute("code", "200");
+				request.setAttribute("message", MESSAGE_DELETE_200);
+			} else {
+				request.setAttribute("code", "400");
+				request.setAttribute("message", MESSAGE_DELETE_404);
+			}
+		} else {
+			request.setAttribute("code", "400");
+			request.setAttribute("message", MESSAGE_DELETE_ATTRIBUTE_MISSING);
 		}
+		
 	}
 
 	@Override
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		System.out.println("BidulePOST !!");
-		System.out.println(request.getParameter("idAlbum"));
-		// modele.Album album = (modele.Album) request.getAttribute("album");
-		// if(album != null){
-		//
-		// DAOFactory.getInstance().getAlbumDao().create(album);
-		// }
+		// Récupére les paramètre de l'album qui doit être créé
+		String titre = request.getParameter("titre");
+		
+		if(titre != ""){
+			Date date = new Date();
+			Personne personne = ((modele.Utilisateur)request.getSession().getAttribute(Connexion.ATT_USER_SESSION)).getPersonne();
+			modele.Album album = new modele.Album(titre, date, personne);
+			DAOFactory.getInstance().getAlbumDao().create(album);
+			
+			request.setAttribute("code", "200");
+			request.setAttribute("message", MESSAGE_POST_200);
+		} else {
+			request.setAttribute("code", "400");
+			request.setAttribute("message", MESSAGE_POST_ATTRIBUTE_MISSING);
+		}
+		
+		
 	}
 
 	@Override
 	protected void doPut(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		// HashMap parameters = (HashMap)
-		// parseParameters(request.getInputStream());
 		System.out.println("BidulePut !!");
 		System.out.println(request.getParameter("idAlbum"));
-		// System.out.println(parameters.get("idAlbum"));
-		//
-		// modele.Album album = (modele.Album)
-		// DAOFactory.getInstance().getAlbumDao().read((int)
-		// parameters.get("idAlbum"));
-		// System.out.println("Album : " + album);
-		// if(album != null){
-		// DAOFactory.getInstance().getAlbumDao().update(album);
-		// }
 	}
-
+	
+	/**********************************************************************************/
+	/**********************************************************************************/
+	/**********************************************************************************/
+	//TODO Ne pas oublier de faire figurer les changements au niveau du serveur tomcat (accepter le parse du body pour les méthodesput et delete
+	/**********************************************************************************/
+	/**********************************************************************************/
+	/**********************************************************************************/
+	
 //	private Map<String, String> parseParameters(ServletInputStream is) throws IOException {
 //		BufferedReader br = new BufferedReader(new InputStreamReader(is));
 //		String data = br.readLine();
